requirements.txt
streamlit==1.30.0
gtts==2.3.2
speechrecognition==3.8.1
pydub==0.25.1
python-dotenv==1.0.1
langchain-google-genai==0.0.11
litellm==0.3.25
